Create folder components + add overlay on hover to album covers
