+++ code/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import data from './data.json'
-import { Header } from './Header'
-import { AlbumList } from './AlbumList'
+import { Header } from './components/Header'
+import { AlbumList } from './components/AlbumList'
 
 
 console.log(data)
