+++ code/src/components/AlbumList.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Album } from './Album';
 import { Artist } from './Artist';
 import { AlbumTitle } from './AlbumTitle';
-import data from './data.json';
+import data from '../data.json';
 
 export const AlbumList = () => {
     return ( 
@@ -10,9 +10,9 @@ export const AlbumList = () => {
         <div className="album-container">
             {data.albums.items.map((item) => (
                 <div className="album-cover">
-                    <Album key={item.id} item={item}/>
-                    <Artist key={item.artists} item={item}/>
+                    <Album key={item.image} item={item}/>
                     <AlbumTitle key={item.name} item={item}/>
+                    <Artist key={item.artists} item={item}/>
                 </div>
             ))}
         </div>
