+++ code/src/components/Artist.js
