+++ code/src/index.css
@@ -70,6 +70,18 @@ header hr {
   display: block;
 }
 
+.image-overlay {
+  position: absolute;
+  top: 5px;
+  width: 300px;
+  height: 300px;
+  transition: 0.3s ease;
+}
+
+.image-overlay:hover {
+  background-color: #000000a1;
+}
+
 .play-button {
   width: 60px;
   height: 60px;
@@ -87,10 +99,6 @@ header hr {
   width: 20px;
   height: 20px;
   padding-top: 15px
-}
-
-.album-image:hover {
-  opacity: 0.6;
 }
 
 .album-image:hover .icons {
